feat: refactoring page
